+++ tests/infrastructures/ProjectBuilding/SfConsoleHookTest.php
@@ -25,6 +25,8 @@ declare(strict_types=1);
 
 namespace Teknoo\Tests\East\Paas\Infrastructures\ProjectBuilding;
 
+use PHPUnit\Framework\MockObject\MockObject;
+use Teknoo\East\Paas\Infrastructures\ProjectBuilding\Contracts\ProcessFactoryInterface;
 use Teknoo\Recipe\Promise\PromiseInterface;
 use Teknoo\East\Paas\Infrastructures\ProjectBuilding\SfConsoleHook;
 use PHPUnit\Framework\TestCase;
@@ -38,16 +40,31 @@ use Symfony\Component\Process\Process;
  */
 class SfConsoleHookTest extends TestCase
 {
+    public function createMock(string $originalClassName): MockObject
+    {
+        return parent::createMock($originalClassName);
+    }
+
     public function buildHook(bool $success = true): SfConsoleHook
     {
         return new SfConsoleHook(
             __DIR__ . '/../../../sfConsole.phar',
-            function () use ($success) {
-                $process = $this->createMock(Process::class);
-                $process->expects(self::any())->method('isSuccessful')->willReturn($success);
+            10,
+            new class($success, $this) implements ProcessFactoryInterface {
+                public function __construct(
+                    private bool $success,
+                    private SfConsoleHookTest $test,
+                ) {
+                }
 
-                return $process;
-            }
+                public function __invoke(array $command, string $cwd, float $timeout): Process
+                {
+                    $process = $this->test->createMock(Process::class);
+                    $process->expects(SfConsoleHookTest::any())->method('isSuccessful')->willReturn($this->success);
+
+                    return $process;
+                }
+            },
         );
     }
 
@@ -179,23 +196,6 @@ class SfConsoleHookTest extends TestCase
         self::assertInstanceOf(
             SfConsoleHook::class,
             $this->buildHook()->setOptions(['action' => 'install', 'arguments' => ['prefer-install']], $promise)
-        );
-    }
-
-    public function testRunNotSfProcess()
-    {
-        $hook = new SfConsoleHook(
-            __DIR__ . '/../../../sfConsole.phar',
-            static fn() => new \stdClass()
-        );
-
-        $promise = $this->createMock(PromiseInterface::class);
-        $promise->expects(self::never())->method('success');
-        $promise->expects(self::once())->method('fail');
-
-        self::assertInstanceOf(
-            SfConsoleHook::class,
-            $hook->run($promise)
         );
     }